+++ src/EventBus.js
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+
+/**
+* Event Bus
+*/
+export const EventBus = new Vue({
+
+});