+++ src/components/Detail.vue
@@ -279,7 +279,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
     .flex.xs12{
       cursor: pointer;   
     }