+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * Transitions + Animations
 * Logo resize
-* 
+* Search realtime
+* IndexedDB